Extract scope data to reuse it to function scanner
